pyproject.toml
[project]
name = "pam-oauth2"
version = "0.1.0"
description = "PAM module for OAuth2/OIDC JWT authentication"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "PyJWT[crypto]>=2.8.0",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "responses>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pam_oauth2"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
